fix(control): disable heading hold when no yaw input to prevent drift
- Only enable heading hold (angle + rate dual-loop) when there's active
  yaw joystick input (threshold: 0.05 rad/s)
- When no yaw input, only apply rate damping to prevent unwanted rotation
- Reset heading integrators when not in use to prevent gyro zero-bias
  accumulation causing continuous rotation
- Add had_yaw_input_ state tracking

This fixes the continuous spinning issue caused by heading_current_
integrating gyro zero-bias even when no yaw command was given.

+++ src/hardware/imu_mpu6050_hal.cpp
@@ -142,9 +142,10 @@ bool MPU6050_HAL::read(ImuData& out) {
   out.gx = gx;
   out.gy = gy;
   out.gz = gz;
-  out.pitch = pitch_ * (180.0f / M_PI);  // Convert to degrees
-  out.roll = roll_ * (180.0f / M_PI);
-  out.yaw = yaw_ * (180.0f / M_PI);
+  // Convert to degrees (M_PI is available in Arduino, cast to float)
+  out.pitch = pitch_ * (180.0f / (float)M_PI);
+  out.roll = roll_ * (180.0f / (float)M_PI);
+  out.yaw = yaw_ * (180.0f / (float)M_PI);
   out.timestamp_us = now_us;
   out.valid = true;
